Use v-for and custom component instead of hardcoded divs
Instead of repeating div's like 
```html
<div class="item">
                <p>1</p>
</div>
```
we can use `v-for` to automatically repeat them for each movie.
```html
<div class="item" v-for="movie in movies">
                <p>1</p>
</div>
```

That will create a `div` for each movie in our `js` file.

However, we don't just want a `div` with a `p` tag inside. We want to show the image and title of the movie and everything.
So something like this:
```html
<div class="item" v-for="movie in movies">
                <img src="test.jpg">
                <p>Movie Title</p>
</div>
```

However, since it's such a common part of our website, we can move that `div`, `img` and `p` tag into it's own file and make a component from it.
Which is what we see inside the `movie-tile.js` file.

So then you end up with the suggested change.